requirements.txt
Flask
pytesseract
Pillow
PyMuPDF
werkzeug
scikit-learn